Refactor BrowserType and Allocator
+ This commit makes Allocator a part of BrowserType as it became an
  unnecessary data structure. There is now BrowserType.Allocate.
+ BrowserType.ExecutablePath now returns the expected browwer path only
  if it didn't cache it before.
+ The Allocate method became so complex (cyclomatic complexity>13).
  The execute method takes some of its responsibilities for command
  execution.
+ Makes some methods functions instead.
